end-of-day-historical-stock-market-data-api

+++ api/api_get_eodhistoricaldata_data.php
@@ -9,18 +9,21 @@
 
     // Get query parameters
     $crypto = $_GET["crypto"];
+    $duration = $_GET["duration"];
 
 // ********************** Get the historical data from API response and store data in CSV file **********************
 
     // Setting API URL
-    $remote_file_name = "https://eodhistoricaldata.com/api/intraday/".$crypto."?api_token=90j6w31n68ka2.91057371";
+    $remote_file_name = "https://eodhistoricaldata.com/api/eod/".$crypto."?api_token=90j6w31n68ka2.91057371&period=".$duration."";
     // echo $remote_file_name."<br>";
 
     //setting file name to save
-    $local_csv_file_name = "../data/historical/".$crypto."/".$crypto."-data.csv"; 
-
-    //Delete the existing file so that we can store new file.
-    unlink($local_csv_file_name); 
+    $local_csv_file_name = "../data/".$duration."/".$crypto."/".$crypto."-data.csv"; 
+    
+    if(file_exists($local_csv_file_name)){
+        //Delete the existing file so that we can store new file.
+        unlink($local_csv_file_name); 
+    }
 
     // Call API
     $ch = curl_init();
@@ -37,5 +40,11 @@
     fclose($file);
 
 // ******************** END - Get the historical data from API response and store data in CSV file ********************
+
+// Output
+header('Content-Type: application/json; charset=utf-8');
+
+echo("Success!");
+
 
 ?>
